# GH-37873: [C++][Parquet] DELTA_BYTE_ARRAY: avoid copying data when possible