### Rationale for this change

When decoding DELTA_BYTE_ARRAY data, if the prefix (respectively suffix) is empty, we don't need to recreate the original string by copying the data into a new buffer, we can just point to the existing suffix (respectively suffix).

### What changes are included in this PR?

Avoid spurious memory copies in the DELTA_BYTE_ARRAY decoder (also reducing the memory footprint when decoding).

Benchmark numbers show that decoding can be up to 2x faster.

### Are these changes tested?

Yes, already tested.

### Are there any user-facing changes?

No.

* Closes: #37873