@mapleFU 2023-09-26
@pitrou Here when all `prefix != 0 && postfix != 0`, it might get slower because CPU Hazard, but it will be faster in some cases