@pitrou 2023-10-03
Updated benchmarks here (AMD Zen 2, Ubuntu 22.04):

* before:
```
-----------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                     Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------------------------------------------------------
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:512/prefixed-percent:10           5018 ns         5017 ns       100000 bytes_per_second=389.09M/s compression_ratio=0.700959 items_per_second=102.047M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:512/prefixed-percent:90           4808 ns         4808 ns       134657 bytes_per_second=427.868M/s compression_ratio=0.765799 items_per_second=106.495M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:512/prefixed-percent:99           4756 ns         4755 ns       135610 bytes_per_second=420.982M/s compression_ratio=0.759524 items_per_second=107.676M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:2048/prefixed-percent:10         21241 ns        21238 ns        33959 bytes_per_second=359.857M/s compression_ratio=0.695357 items_per_second=96.4296M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:2048/prefixed-percent:90         21215 ns        21211 ns        33216 bytes_per_second=368.188M/s compression_ratio=0.753808 items_per_second=96.5536M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:2048/prefixed-percent:99         20278 ns        20274 ns        33480 bytes_per_second=386.568M/s compression_ratio=0.762051 items_per_second=101.016M/s

BM_DeltaDecodingByteArray/max-string-length:64/batch-size:512/prefixed-percent:10          4737 ns         4736 ns       148976 bytes_per_second=3.20831G/s compression_ratio=0.933408 items_per_second=108.115M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:512/prefixed-percent:90          5002 ns         5001 ns       136927 bytes_per_second=3.08213G/s compression_ratio=1.19315 items_per_second=102.369M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:512/prefixed-percent:99          4987 ns         4986 ns       139231 bytes_per_second=3.01671G/s compression_ratio=1.21214 items_per_second=102.684M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:2048/prefixed-percent:10        19989 ns        19986 ns        35066 bytes_per_second=3.02942G/s compression_ratio=0.930448 items_per_second=102.472M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:2048/prefixed-percent:90        21737 ns        21732 ns        31170 bytes_per_second=2.81764G/s compression_ratio=1.18361 items_per_second=94.2395M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:2048/prefixed-percent:99        22059 ns        22054 ns        30975 bytes_per_second=2.77557G/s compression_ratio=1.22369 items_per_second=92.8648M/s

BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:512/prefixed-percent:10        8545 ns         8543 ns        79054 bytes_per_second=27.8606G/s compression_ratio=1.0309 items_per_second=59.9351M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:512/prefixed-percent:90        9102 ns         9100 ns        77461 bytes_per_second=25.8281G/s compression_ratio=1.43071 items_per_second=56.2662M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:512/prefixed-percent:99        9429 ns         9426 ns        73866 bytes_per_second=26.8462G/s compression_ratio=1.54352 items_per_second=54.3178M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:2048/prefixed-percent:10      34260 ns        34255 ns        20326 bytes_per_second=28.1638G/s compression_ratio=1.02877 items_per_second=59.7862M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:2048/prefixed-percent:90      38045 ns        38038 ns        18457 bytes_per_second=25.7488G/s compression_ratio=1.41325 items_per_second=53.8402M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:2048/prefixed-percent:99      38601 ns        38594 ns        18100 bytes_per_second=26.0021G/s compression_ratio=1.50832 items_per_second=53.0648M/s
```

* after:
```
-----------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                                                                     Time             CPU   Iterations UserCounters...
-----------------------------------------------------------------------------------------------------------------------------------------------
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:512/prefixed-percent:10           2701 ns         2701 ns       245508 bytes_per_second=722.842M/s compression_ratio=0.700959 items_per_second=189.581M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:512/prefixed-percent:90           4227 ns         4226 ns       153065 bytes_per_second=486.726M/s compression_ratio=0.765799 items_per_second=121.145M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:512/prefixed-percent:99           4173 ns         4172 ns       152447 bytes_per_second=479.758M/s compression_ratio=0.759524 items_per_second=122.71M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:2048/prefixed-percent:10         10383 ns        10381 ns        66118 bytes_per_second=736.207M/s compression_ratio=0.695357 items_per_second=197.279M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:2048/prefixed-percent:90         17619 ns        17615 ns        37814 bytes_per_second=443.344M/s compression_ratio=0.753808 items_per_second=116.262M/s
BM_DeltaDecodingByteArray/max-string-length:8/batch-size:2048/prefixed-percent:99         18600 ns        18597 ns        35834 bytes_per_second=421.419M/s compression_ratio=0.762051 items_per_second=110.123M/s

BM_DeltaDecodingByteArray/max-string-length:64/batch-size:512/prefixed-percent:10          2626 ns         2625 ns       252886 bytes_per_second=5.78763G/s compression_ratio=0.933408 items_per_second=195.034M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:512/prefixed-percent:90          5100 ns         5099 ns       139164 bytes_per_second=3.02322G/s compression_ratio=1.19315 items_per_second=100.413M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:512/prefixed-percent:99          5138 ns         5137 ns       131273 bytes_per_second=2.92809G/s compression_ratio=1.21214 items_per_second=99.6678M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:2048/prefixed-percent:10        10647 ns        10646 ns        65412 bytes_per_second=5.68743G/s compression_ratio=0.930448 items_per_second=192.38M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:2048/prefixed-percent:90        20188 ns        20184 ns        32917 bytes_per_second=3.03371G/s compression_ratio=1.18361 items_per_second=101.466M/s
BM_DeltaDecodingByteArray/max-string-length:64/batch-size:2048/prefixed-percent:99        21673 ns        21669 ns        29778 bytes_per_second=2.82478G/s compression_ratio=1.22369 items_per_second=94.5112M/s

BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:512/prefixed-percent:10        3004 ns         3004 ns       231770 bytes_per_second=79.2403G/s compression_ratio=1.0309 items_per_second=170.465M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:512/prefixed-percent:90        8603 ns         8602 ns        81008 bytes_per_second=27.3233G/s compression_ratio=1.43071 items_per_second=59.5234M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:512/prefixed-percent:99        9428 ns         9426 ns        73990 bytes_per_second=26.8451G/s compression_ratio=1.54352 items_per_second=54.3156M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:2048/prefixed-percent:10      12057 ns        12055 ns        58266 bytes_per_second=80.0331G/s compression_ratio=1.02877 items_per_second=169.895M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:2048/prefixed-percent:90      36052 ns        36046 ns        19424 bytes_per_second=27.1721G/s compression_ratio=1.41325 items_per_second=56.8163M/s
BM_DeltaDecodingByteArray/max-string-length:1024/batch-size:2048/prefixed-percent:99      39117 ns        39111 ns        17756 bytes_per_second=25.6587G/s compression_ratio=1.50832 items_per_second=52.364M/s
```